requirements.txt
configparser
pymysql
boto3
psycopg2
mysql-replication